Refactor imports and enhance type safety in ProductList and ProtectedRouteWrapper components
