+++ src/App.tsx
@@ -1,5 +1,4 @@
 import { Outlet } from "react-router"
-import { Toaster } from "./components/ui/toaster"
 
 function App() {
